Fix compile error due to enum changes
Merge request !23 changed to strongly typed enums,
so we now need to explicitly cast the output in the
Python getters.

+++ pcbnew/swig/pcbnew_scripting_helpers.cpp
@@ -171,7 +171,7 @@ void UpdateUserInterface()
 int GetUserUnits()
 {
     if( s_PcbEditFrame )
-        return s_PcbEditFrame->GetUserUnits();
+        return static_cast<int>( s_PcbEditFrame->GetUserUnits() );
 
     return -1;
 }
